refactor: split flow components
