+++ package.json
@@ -8,7 +8,7 @@
     "build": "tsc --project tsconfig.build.json && vite build",
     "test": "vitest",
     "test:watch": "vitest --watch",
-    "test:coverage": "vitest --coverage",
+    "test:coverage": "node scripts/test-coverage.mjs",
     "test:ui": "vitest --ui",
     "test:integration": "vitest --config vitest.integration.config.ts",
     "test:e2e": "vitest --config vitest.e2e.config.ts",
